Fixed bugs causing getAllGroceryStores to make duplicate queries.
1: There was some space between query buffers due to how the buffer were
generate. Switching the number of segments in the buffer to a number
divisble by 6 fixed this issue.

2: Unsearched buffer query sometimes returned the edge of the query area
as unsearched when it shouldn't. I'm not sure exactly what caused this
issue but, it was resolved by implementing selectUnsearchedBuffer in SQL
dirrectly instead of in terms of selectSearchedBuffer.

Tests also added for these cases.